+++ src/app/post/create/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 import Link from 'next/link'
 import { useSelector } from 'react-redux'
-import { BackNavbar } from '@/app/_components/Common/Navbar/Back'
+import { BackNavbar } from '@/app/_components/Common/Navbar/BackNavigationBar'
 import CreatePost from '@/app/post/create/_components/CreatePost'
 import { selectSelectedShop } from '@/store/features/shopSlice'
